Fix tn25 crud product and update image profile (#32)
* fix-TN25-CRUDProduct:: add search Product

* fix-TN25-CRUDProduct updateImage and CRUD Product

+++ mvc/controllers/AdminProductController.php
@@ -3,7 +3,8 @@ require_once dirname(__DIR__) . "/models/Product.php";
 
 
 
-class AdminProductController{
+class AdminProductController
+{
     function showProduct()
     {
         $productModel = new Product();
@@ -39,39 +40,86 @@ class AdminProductController{
     }
     function updateProduct1()
     {
-        $productModel = new Product();
+        $name = "";
+        $description = "";
+        $categories = "";
+        $image_name = "";
+        $qty = "";
+        $price = "";
+        $image_url = "";
+
+        $name_error = "";
+        $description_error = "";
+        $categories_error = "";
+        $image_name_error = "";
+        $qty_error = "";
+        $price_error = "";
+        $image_url_error = "";
+
         if (isset($_POST['updateProduct'])) {
             $id = htmlspecialchars($_POST['id']);
-            $name = htmlspecialchars($_POST['PName']);
-            $description = htmlspecialchars($_POST['PDescription']);
-            $categories = htmlspecialchars($_POST['PCategories']);
-            $image_name = htmlspecialchars($_POST['PImage_name']);
-            $qty = intval($_POST['PQty']);
-            $price = floatval($_POST['PPrice']);
-            $image_url = $_ENV['ROOT_URL'] . '/public/image/' . basename($_FILES["PImage_url"]["name"]);
+            $name = isset($_POST['PName']) ? htmlspecialchars($_POST['PName']) : "";
+            if (empty($name)) {
+                $name_error = "Please enter product name";
+            }
+            $description = isset($_POST['PDescription']) ? htmlspecialchars($_POST['PDescription']) : "";
+            if (empty($description)) {
+                $description_error = "Please description";
+            }
+            $categories = isset($_POST['PCategories']) ? htmlspecialchars($_POST['PCategories']) : "";
+            if (empty($categories)) {
+                $categories_error = "Please enter categories";
+            }
+            $image_name = isset($_POST['PImage_name']) ? htmlspecialchars($_POST['PImage_name']) : "";
+            if (empty($image_name)) {
+                $image_name_error = "Please enter image name";
+            }
+
+            $qty = isset(($_POST['PQty'])) ? htmlspecialchars(intval($_POST['PQty'])) : "";
+            if (empty($qty)) {
+                $qty_error = "Please enter quantity";
+            } elseif ($qty < 0) {
+                $qty_error = "Invalid quantity";
+            }
+            $price = isset(($_POST['PPrice'])) ? htmlspecialchars(floatval($_POST['PPrice'])) : "";
+            if (empty($price)) {
+                $price_error = "Please enter price";
+            } elseif ($price < 0) {
+                $price_error = "Invalid price";
+            }
+            $image_url = isset($_POST['PImage_url']) ? htmlspecialchars($_POST['PImage_url']) : "";
             if (isset($_FILES["PImage_url"]["tmp_name"]) && !empty($_FILES["PImage_url"]["tmp_name"])) {
 
                 $imageFileType = strtolower(pathinfo($_FILES["PImage_url"]["name"], PATHINFO_EXTENSION));
 
                 if ($imageFileType != "jpg" && $imageFileType != "png" && $imageFileType != "jpeg" && $imageFileType != "gif") {
+                    $image_url_error = "Invalid photo";
+                } else {
 
-                    echo "Only JPG, JPEG, PNG, and GIF files are allowed.";
-                    return;
+                    $target_dir = dirname(dirname(__DIR__)) . '/public/image/';
+                    $target_file = $target_dir . basename($_FILES["PImage_url"]["name"]);
+                    move_uploaded_file($_FILES["PImage_url"]["tmp_name"], $target_file);
+                    if (!file_exists($target_file)) {
+                        move_uploaded_file($_FILES["PImage_url"]["tmp_name"], $target_file);
+                    }
+                    $image_url = $_ENV['ROOT_URL'] . '/public/image/' . basename($_FILES["PImage_url"]["name"]);
                 }
-                $target_dir = dirname(dirname(__DIR__)) . '/public/image/';
-                $target_file = $target_dir . basename($_FILES["PImage_url"]["name"]);
-
-                move_uploaded_file($_FILES["PImage_url"]["tmp_name"], $target_file);
-            } else {
-
-                echo "Please choose an image.";
-                return;
             }
-            $updateProduct = $productModel->updateProductQ($id, $name, $description, $categories, $image_name, $image_url, $qty, $price);
+        }
+
+        if (empty($name_error) && empty($description_error) && empty($categories_error) && empty($image_name_error) && empty($qty_error) && empty($price_error) && empty($image_url_error)) {
+            $productModel = new Product();
+
+            $updateProduct = $productModel->updateProduct($id, $name, $description, $categories, $image_name, $image_url, $qty, $price);
             if ($updateProduct == true) {
                 header('location: showProduct');
                 exit();
             }
+        } else {
+            $productModel = new Product();
+            $id = $_POST['id'];
+            $productOne = $productModel->getOne($id);
+            view("admin/products/update", compact('productOne', 'name_error', 'description_error', 'categories_error', 'image_name_error', 'image_url_error', 'qty_error', 'price_error'));
         }
     }
 
@@ -82,16 +130,53 @@ class AdminProductController{
 
     function insertProduct1()
     {
-        $filetypeError = "";
+        $name = "";
+        $description = "";
+        $categories = "";
+        $image_name = "";
+        $qty = "";
+        $price = "";
+        $image_url = "";
 
+        $name_error = "";
+        $description_error = "";
+        $categories_error = "";
+        $image_name_error = "";
+        $qty_error = "";
+        $price_error = "";
+        $image_url_error = "";
         if (isset($_POST['insertProduct1'])) {
-            $productModel = new Product();
-            $name = htmlspecialchars($_POST['PName']);
-            $description = htmlspecialchars($_POST['PDescription']);
-            $categories = htmlspecialchars($_POST['PCategories']);
-            $image_name = htmlspecialchars($_POST['PImage_name']);
-            $qty = intval($_POST['PQty']);
-            $price = floatval($_POST['PPrice']);
+
+            $name = isset($_POST['PName']) ? htmlspecialchars($_POST['PName']) : "";
+            if (empty($name)) {
+                $name_error = "Please enter product name";
+            }
+            $description = isset($_POST['PDescription']) ? htmlspecialchars($_POST['PDescription']) : "";
+            if (empty($description)) {
+                $description_error = "Please description";
+            }
+            $categories = isset($_POST['PCategories']) ? htmlspecialchars($_POST['PCategories']) : "";
+            if (empty($categories)) {
+                $categories_error = "Please enter categories";
+            }
+            $image_name = isset($_POST['PImage_name']) ? htmlspecialchars($_POST['PImage_name']) : "";
+            if (empty($image_name)) {
+                $image_name_error = "Please enter image name";
+            }
+
+            $qty = isset(($_POST['PQty'])) ? htmlspecialchars(intval($_POST['PQty'])) : "";
+            if (empty($qty)) {
+                $qty_error = "Please enter quantity";
+            } elseif ($qty < 0) {
+                $qty_error = "Invalid quantity";
+            }
+            $price = isset(($_POST['PPrice'])) ? htmlspecialchars(floatval($_POST['PPrice'])) : "";
+            if (empty($price)) {
+                $price_error = "Please enter price";
+            } elseif ($price < 0) {
+                $price_error = "Invalid price";
+            }
+
             $image_url = $_ENV['ROOT_URL'] . '/public/image/' . basename($_FILES["PImage_url"]["name"]);
 
             if (isset($_FILES["PImage_url"]["tmp_name"]) && !empty($_FILES["PImage_url"]["tmp_name"])) {
@@ -100,27 +185,49 @@ class AdminProductController{
 
                 if ($imageFileType != "jpg" && $imageFileType != "png" && $imageFileType != "jpeg" && $imageFileType != "gif") {
 
-                    $filetypeError = "you should choose photo";
-                    return  $filetypeError;
+                    $image_url_error = "Invalid photo";
+                } else {
+
+                    $target_dir = dirname(dirname(__DIR__)) . '/public/image/';
+                    $target_file = $target_dir . basename($_FILES["PImage_url"]["name"]);
+                    move_uploaded_file($_FILES["PImage_url"]["tmp_name"], $target_file);
+                    if (!file_exists($target_file)) {
+                        move_uploaded_file($_FILES["PImage_url"]["tmp_name"], $target_file);
+                    }
+                    $image_url = $_ENV['ROOT_URL'] . '/public/image/' . basename($_FILES["PImage_url"]["name"]);
                 }
-
-                $target_dir = dirname(dirname(__DIR__)) . '/public/image/';
-                $target_file = $target_dir . basename($_FILES["PImage_url"]["name"]);
-
-
-                move_uploaded_file($_FILES["PImage_url"]["tmp_name"], $target_file);
             } else {
-
-                $filetypeError = "you should choose photo";
+                $image_url_error = "Please choose photo";
             }
+        }
 
+        if (empty($name_error) && empty($description_error) && empty($categories_error) && empty($image_name_error) && empty($qty_error) && empty($price_error) && empty($image_url_error)) {
+            $productModel = new Product();
             $insertProduct = $productModel->insertProduct($name, $description, $categories, $image_name, $image_url, $qty, $price);
             if ($insertProduct == true) {
                 header('location: showProduct');
                 exit();
             }
+        } else {
+            view("admin/products/insert", compact('name_error', 'description_error', 'categories_error', 'image_name_error', 'image_url_error', 'qty_error', 'price_error'));
         }
     }
-}
+    public  function search($keyword)
+    {
+        $search_key = "";
+        $keyword = isset($_POST['key']) ? htmlspecialchars($_POST['key']) : '';
+        $productModel = new Product();
+        $searchResult = $productModel->search($keyword);
+        if (!empty($keyword)) {
+            if (empty($searchResult)) {
+                $search_key = "No results ";
+                $search_key = "No results : " . htmlspecialchars($keyword);
 
-?>
+                view('error/error', compact('search_key'));
+            } else {
+                $search_key =  htmlspecialchars($keyword);
+            }
+        }
+        view('admin/products/search', compact('searchResult', 'keyword', 'search_key'));
+    }
+}
